Move logging into utilities
