--- firmware/gameport-adapter/Log.h
@@ -1,30 +1,0 @@
-// This file is part of Necroware's GamePort adapter firmware.
-// Copyright (C) 2021 Necroware
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-#pragma once
-
-#include <Arduino.h>
-#include <stdarg.h>
-#include <stdio.h>
-
-inline void log(const char *fmt, ...) {
-  va_list args;
-  va_start(args, fmt);
-  char buffer[512];
-  vsnprintf(buffer, sizeof(buffer), fmt, args);
-  Serial.println(buffer);
-  va_end(args);
-}
